+++ src/Service/ServiceManagerConfig.php
@@ -11,6 +11,7 @@ namespace Zend\Mvc\Service;
 
 use Zend\EventManager\EventManagerAwareInterface;
 use Zend\EventManager\EventManagerInterface;
+use Zend\EventManager\SharedEventManagerInterface;
 use Zend\ServiceManager\Config;
 use Zend\ServiceManager\ServiceLocatorAwareInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
@@ -94,15 +95,20 @@ class ServiceManagerConfig extends Config
     {
         $this->initializers = [
             'EventManagerAwareInitializer' => function ($instance, ServiceLocatorInterface $serviceLocator) {
-                if ($instance instanceof EventManagerAwareInterface) {
-                    $eventManager = $instance->getEventManager();
-
-                    if ($eventManager instanceof EventManagerInterface) {
-                        $eventManager->setSharedManager($serviceLocator->get('SharedEventManager'));
-                    } else {
-                        $instance->setEventManager($serviceLocator->get('EventManager'));
-                    }
+                if (! $instance instanceof EventManagerAwareInterface) {
+                    return;
                 }
+
+                $eventManager = $instance->getEventManager();
+
+                // If the instance has an EM WITH an SEM composed, do nothing.
+                if ($eventManager instanceof EventManagerInterface
+                    && $eventManager->getSharedManager() instanceof SharedEventManagerInterface
+                ) {
+                    return;
+                }
+
+                $instance->setEventManager($serviceLocator->get('EventManager'));
             },
             'ServiceManagerAwareInitializer' => function ($instance, ServiceLocatorInterface $serviceLocator) {
                 if ($serviceLocator instanceof ServiceManager && $instance instanceof ServiceManagerAwareInterface) {