+++ src/View/Console/InjectNamedConsoleParamsListener.php
@@ -19,7 +19,7 @@ class InjectNamedConsoleParamsListener extends AbstractListenerAggregate
     /**
      * {@inheritDoc}
      */
-    public function attach(Events $events)
+    public function attach(Events $events, $priority = 1)
     {
         $this->listeners[] = $events->attach(MvcEvent::EVENT_DISPATCH, [$this, 'injectNamedParams'], -80);
     }