+++ src/View/Http/DefaultRenderingStrategy.php
@@ -45,7 +45,7 @@ class DefaultRenderingStrategy extends AbstractListenerAggregate
     /**
      * {@inheritDoc}
      */
-    public function attach(EventManagerInterface $events)
+    public function attach(EventManagerInterface $events, $priority = 1)
     {
         $this->listeners[] = $events->attach(MvcEvent::EVENT_RENDER, [$this, 'render'], -10000);
         $this->listeners[] = $events->attach(MvcEvent::EVENT_RENDER_ERROR, [$this, 'render'], -10000);
@@ -108,9 +108,11 @@ class DefaultRenderingStrategy extends AbstractListenerAggregate
 
             $application = $e->getApplication();
             $events      = $application->getEventManager();
-            $e->setError(Application::ERROR_EXCEPTION)
-              ->setParam('exception', $ex);
-            $events->trigger(MvcEvent::EVENT_RENDER_ERROR, $e);
+
+            $e->setError(Application::ERROR_EXCEPTION);
+            $e->setParam('exception', $ex);
+            $e->setName(MvcEvent::EVENT_RENDER_ERROR);
+            $events->triggerEvent($e);
         }
 
         return $response;